# Optimize scheduler cache snapshot to improve scheduling throughput