**What type of PR is this?**
/kind feature

**What this PR does / why we need it**:
The scheduler snapshots its cache at the beginning of each scheduling cycle to have a persistent view of the cluster during the cycle. CPU profiling results shows that the scheduler spends a significant percentage of a scheduling cycle for snapshotting its cache.
This PR is an algorithmic optimization that dramatically improves performance of the cache snapshot logic.
The algorithm works by keeping the cache entries in a doubly linked list where the most recently updated entry is at the head and the rest of the entries are ordered by the recency of the updates. These entries have had a global generation number (before this PR). The snapshotting logic starts traversing the doubly linked list and updates entries in the cache snapshot that are updated after the last snapshot.

Our benchmarks shows the latency of scheduling a pod is improved over 20% from 8.5ms to 6.7ms in a 5000 node cluster when scheduling 10000 pods:

```
BEFORE:
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkScheduling/5000Nodes/1000Pods-12         	   10000	   8516461 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	107.756s

----------------------------------------
AFTER:

pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkScheduling/5000Nodes/1000Pods-12         	   10000	   6740178 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	89.193s
```
**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--  
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note
Optimize scheduler cache snapshot algorithm to improve scheduling throughput.
```

/sig scheduling