@bsalamat 2019-02-21
Our scalability results after this PR are in and available in our [perf dashboard](http://perf-dash.k8s.io/). 
I am glad to announce that average scheduling throughput has reached 100 pods/s in a 5000 node cluster. This was an important milestone for improving scalability of Kubernetes scheduler.

Choose "gce-5000Nodes > Scheduler > SchedulingThroughput" in the perf dashboard to see the results. Latest at the time of this comment is run 314.
Our scalability tests are rate limited at 100 pods/s. So, the scheduler might be able to achieve even higher throughput if the rate limit is lifted.