@wgliang 2019-02-22
It looks great, but I don't know if you noticed that our binding time (Perc99/Perc90) has increased a lot.