@bsalamat 2019-02-23
> It looks great, but I don't know if you noticed that our binding time (Perc99/Perc90) has increased a lot. Choose "gce-5000Nodes > Scheduler > SchedulingLantency->binding" in the perf dashboard.

Thanks for noticing that. Average binding time has not changed, but 90th percentile has increased a lot. I think I know the reason. We have a rate limit of 100 qps to the API server. Scheduler's throughput now exceeds that limit and can schedule more than 100 pods per second. As a result pods wait longer in binding stage and we see an increase in binding latency.